% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_enrichments.R
\name{calculate_enrichments}
\alias{calculate_enrichments}
\title{parallelized enrichment testing for ST data}
\usage{
calculate_enrichments(
  counts,
  db_dataset = "mmusculus_gene_ensembl",
  n.cores = 4
)
}
\arguments{
\item{counts}{non-negative numeric matrix. rows correspond to spots,
columns correspond to genes}

\item{db_dataset}{ensembl dataset to use, default 'mmusculus_gene_ensembl'}

\item{ncores}{number of threads to be used for parallel execution}
}
\value{
list containing an enrichment table for each spot
}
\description{
parallelized enrichment testing for ST data
}
