% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spot_enrichments.R
\name{calculate_spot_enrichments}
\alias{calculate_spot_enrichments}
\title{create an enrichment matrix for a given ST count matrix}
\usage{
calculate_spot_enrichments(
  counts,
  n.cores = 4,
  db_dataset = "mmusculus_gene_ensembl"
)
}
\arguments{
\item{counts}{non-negative numeric matrix. rows correspond to spots,
columns correspond to genes}

\item{db_dataset}{ensembl dataset to use, default 'mmusculus_gene_ensembl'}

\item{ncores}{number of threads to be used for parallel execution}
}
\value{
list with two entries: \cr
\enumerate{
\item enrichment.list: list containing an enrichment table for each spot in the count matrix\cr
\item term.matrix: numeric matrix, terms x spots containing the p-value for each term in each spot
}
}
\description{
create an enrichment matrix for a given ST count matrix
}
