% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ids.R
\name{clean_ids}
\alias{clean_ids}
\title{This function selects spots based on connectivity. Based on the assumption
that only spots with a certain number of connected neighbours are of interest,
spots that are not sufficiently connected to the main part of the tissue are removed.}
\usage{
clean_ids(ids)
}
\arguments{
\item{ids}{ids data frame as produced by process_input(), reduced to rows corresponding to the
spots to be retained based on image analysis (remove_background)}
}
\value{
new ids data frame
}
\description{
This function selects spots based on connectivity. Based on the assumption
that only spots with a certain number of connected neighbours are of interest,
spots that are not sufficiently connected to the main part of the tissue are removed.
}
