% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_counts.R
\name{cluster_counts}
\alias{cluster_counts}
\title{cluster counts matrix based using louvain clustering based on UMAP and PCA}
\usage{
cluster_counts(
  counts,
  ids = NULL,
  pca.fraction = 0.8,
  umap.metric = "cosine",
  ...
)
}
\arguments{
\item{counts}{non-negative numeric matrix containing gene counts,
rows correspond to spots, columns correspond to genes}

\item{ids}{data frame or matrix assigning spatial coordinates to the spots, used
for clustering in combination with the UMAP embedding, optional}

\item{pca.fraction}{numeric, 0 < pca.fraction < 1. Fraction of variance to be explained by
principal components taken into account. default 0.8}

\item{umap.metric}{string, metric passed to uwot::umap for umap embedding. default 'cosine'}

\item{..., }{parameters passed to uwot::umap function for fine-tuning}
}
\value{
list containing two elements:\cr
\enumerate{
\item clustering - numeric clustering vector, contains clusters for all spots in the same order as rownames(counts)\cr
\item umap.plot - ggplot object, UMAP plot of count matrix coloured by clustering
\item spatial.plot - ggplot object, spatial representation of clustering; only if parameter
}
}
\description{
cluster counts matrix based using louvain clustering based on UMAP and PCA
}
