% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spots.R
\name{cluster_spots}
\alias{cluster_spots}
\title{cluster spots using enrichment analysis, kmeans and hierarchical clustering}
\usage{
cluster_spots(
  counts,
  spot.enrichments = NULL,
  term.cutoff = 10,
  spot.cutoff = 1,
  n.cluster.spots = 6,
  n.cluster.terms = 6,
  n.cores = 4,
  verbose = TRUE,
  output_folder = NULL,
  method = "normal"
)
}
\arguments{
\item{counts}{non-negative numeric matrix. rows correspond to spots,
columns correspond to genes}

\item{spot.enrichments}{list with two entries as returned by calculate_spot_enrichments(): \cr
\enumerate{
\item enrichment.list: list containing an enrichment table for each spot in the count matrix\cr
\item term.matrix: numeric matrix, terms x spots containing the p-value for each term in each spot
}}

\item{term.cutoff}{numeric > 0, minimum number of spots a term needs to be enriched in in order to be
considered in the analysis}

\item{spot.cutoff}{numeric > 0, minimum number of terms that have to be enriched in a spot for it be
considered in the analysis}

\item{n.cluster.spots}{numeric > 0, number of clusters the spots should be grouped in}

\item{n.cluster.terms}{numeric > 0, number of clusters the terms should be grouped in}

\item{verbose}{logical}

\item{output_folder}{string specifying the directory in which plots generated by this function should be stored; default NULL, no plots generated}

\item{method}{string specifying the clustering method to be used. default: 'normal'}

\item{ncores}{number of threads to be used for parallel execution}
}
\value{
list containing five elements\cr
\enumerate{
\item spot.enrichments: list as returned by calculate_spot_enrichments()\cr
\item enrichment.mat: numeric matrix, terms x spots containing the p-value for each term in each spot\cr
\item spot.clustering: clustering information for spots in enrichment.mat
\item term.clustering: clustering information for terms in enrichment.mat
\item term.clusters: list containing for each term cluster a vector of terms belonging to that cluster
}
}
\description{
cluster spots using enrichment analysis, kmeans and hierarchical clustering
}
