% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_error.R
\name{coef_error}
\alias{coef_error}
\title{Helper function for error handling}
\usage{
coef_error(lasso, lambda)
}
\arguments{
\item{lasso}{lasso model object you want to evaluate}

\item{lambda}{lambda value at which you are trying to evaluate}
}
\value{
returns the nearest possible evaluation to your lambda
}
\description{
Helper function for error handling
}
\details{
This function catches errors due to truncation of the lasso solution path and returns output for the best available lambda
}
