% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_compute_poisson_potential.R
\name{compute_poisson_potential}
\alias{compute_poisson_potential}
\title{Compute the attractivity potential on a grid from balance value on polygon following the method of Tobler.}
\usage{
compute_poisson_potential(
  xsf,
  varname = "delta",
  method = "jacobi",
  nb_it = 1000,
  cellsize = sqrt((sf::st_bbox(xsf)[3] - sf::st_bbox(xsf)[1]) * (sf::st_bbox(xsf)[4] -
    sf::st_bbox(xsf)[2])/2500)
)
}
\arguments{
\item{xsf}{an sf data.frame of polygons}

\item{varname}{name of the feature where the balance of each polygons are stored}

\item{method}{resolution method for poisson equation either jacobi or solve}

\item{nb_it}{number of iteration for jacobi resolution}

\item{cellsize}{size of the}
}
\value{
an sf data.frame with
}
\description{
Compute the attractivity potential on a grid from balance value on polygon following the method of Tobler.
TTT equivalent of the potflow program of Waldo Tobler
POTENTIAL FIELD FROM DATA GIVEN BY POLYGONS CONVERTED
TO A LATTICE USING THE POLYGRID PROGRAM.
COMPUTED AS THE SOLUTION TO THE POISSON EQUATION
WITH BOUNDARY CONDITION DZ/DN=0.
SUMMER 1978, REVISED, CONVERTED TO BASIC, SPRING 1994. Rev 6/8/98
\code{poisson.potential} returns an sf
}
\examples{
data(dollars)
dollars$polygones$delta = rowSums(dollars$OD)-colSums(dollars$OD)
pot=compute_poisson_potential(dollars$polygones)
}
