% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test.R
\name{enrichment_test}
\alias{enrichment_test}
\title{perform enrichment testing using topGO and fisher's exact test for a given gene list}
\usage{
enrichment_test(
  all.genes,
  genelist,
  db,
  go_ids = NULL,
  sig.level = 0.05,
  max.terms = 5000,
  node.size = 10
)
}
\arguments{
\item{all.genes}{character vector, the reference gene set. for example all genes in the experiment}

\item{genelist}{character vector, list of interesting genes}

\item{db}{biomaRt database for the data}

\item{go_ids}{table assigning GO ids to gene names, obtained using getBM}

\item{sig.level}{numeric, 0 < sig.level < 1. Significance cutoff, p-values above this threshold are discarded}

\item{max.terms}{numeric > 0, maximum number of terms to be considered}

\item{node.size}{ignore GO nodes with less than node.size annotated genes}
}
\value{
table containing information about each significant GO term
}
\description{
perform enrichment testing using topGO and fisher's exact test for a given gene list
}
