% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_ids.R
\name{fill_ids}
\alias{fill_ids}
\title{fill the barcode data frame with artificial entries for missing coordinates}
\usage{
fill_ids(ids)
}
\arguments{
\item{ids}{barcode data frame assigning spatial coordinates to all spots; column names must be 'X' and 'Y'}
}
\value{
full ids data frame
}
\description{
fill the barcode data frame with artificial entries for missing coordinates
}
