% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spot_enrichments.R
\name{filter_spot_enrichments}
\alias{filter_spot_enrichments}
\title{this function filters the enrichment matrix for interesting terms based on number of spots a term is enriched in and
the terms specificity to one cluster}
\usage{
filter_spot_enrichments(
  enrichment.mat,
  term.names,
  clustering = NULL,
  spot.threshold = 10,
  specificity.threshold = 0.75
)
}
\arguments{
\item{enrichment.mat}{numeric matrix of spot enrichments containing p-values (GO terms x spots)
as produced by test_spot_enrichment()}

\item{term.names}{character vector, full names of GO ids that are the rownames of enrichment.mat}

\item{clustering}{numeric vector assigning each spot (in the same order as in matrix) to a cluster}

\item{spot.threshold}{numeric > 0, number of spots a GO term needs to be enriched in}

\item{specificity.theshold}{numeric, 0<x<1; fraction of spots that a Term is enriched in that need
to be in the same cluster for the Term to be deemed interesting}
}
\value{
list containing two enries:\cr
\enumerate{
\item enrichment.mat - filtered matrix contining p-values of enrichments test for
each spot and GO term (terms x spots) \cr
\item term.names - character vector containing full names of the GO ids in the enrichment matrix
}
}
\description{
this function filters the enrichment matrix for interesting terms based on number of spots a term is enriched in and
the terms specificity to one cluster
}
