% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso_complete_fixed_gamma.R
\name{fused_lasso_complete_fixed_gamma}
\alias{fused_lasso_complete_fixed_gamma}
\title{Building sparse fused lasso models for given genes from their count data}
\usage{
fused_lasso_complete_fixed_gamma(
  counts_matrix,
  ids_table,
  name,
  output_folder,
  gamma = 1
)
}
\arguments{
\item{counts_matrix}{a matrix with gene expression levels across genes (columns) and associated spots (rows)}

\item{ids_table}{table  associating each rowname of the counts_matrix with an x and y coordinate}

\item{name}{name of the object to save to}

\item{output_folder}{folder to save object in}

\item{gamma}{ratio of penalization between sparsity and continuity (default 1, high = sparse)}
}
\value{
list containing three elements\cr
\enumerate{
\item counts - matrix containing the fits (Y) for each gene's model
\item BICs - vector containing the BIC for each model
\item lls - vector containing the likelihood for each model
}
}
\description{
Building sparse fused lasso models for given genes from their count data
}
\details{
This function will build a sparse fused lasso for each input gene from it's expression vector and the associated spatial coordinates and save the resulting models to an .Rdata object
}
