% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_view.R
\name{get_tt}
\alias{get_tt}
\title{Get a table of tracked time within a range of dates}
\usage{
get_tt(s = NULL, e = NULL, last_n_days = NULL)
}
\arguments{
\item{s}{minimum day for which tasks are to be
collected. Date must be in ISO format:
\code{YYYY-MM-DD}.}

\item{e}{maximum day for which tasks are to be
collected. Date must be in ISO format:
\code{YYYY-MM-DD}.}

\item{last_n_days}{a number of days back from the
current day for which tasks are to be collected.}
}
\description{
Get a table of tracked time (as a \code{tbl_df}
object). The date range or number of days back
from the present day can be specified to filter
the table returned. Otherwise, all tracked time
records are provided.
}
\examples{
\dontrun{
# Show all tracked tasks as a tibble
get_tt()
#> A tibble: 3 x 8
#>   task                    s_d        s_t   e_d    e_t   lbl   proj   info
#>   <chr>                   <chr>      <chr> <chr>  <chr> <chr> <chr>  <chr>
#> 1 Description of the task 2018-05-24 11:30 2018-… 12:00 label proje… Any a…
#> 2 Description of the task 2018-05-24 12:00 2018-… 12:30 label proje… Any a…
#> 3 Description of the task 2018-05-24 12:30 2018-… 13:00 label proje… Any a…
}
}
