% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt.R
\name{knit_print.ttt}
\alias{knit_print.ttt}
\title{Method for printing in a \code{knitr} context}
\usage{
\method{knit_print}{ttt}(x, ..., theme = getOption("ttt.theme"))
}
\arguments{
\item{x}{An object returned by \code{\link{ttt}}.}

\item{...}{Further arguments passed on to \code{knitr::knit_print}.}

\item{theme}{A theme (either "default" or "booktabs").}
}
\description{
Method for printing in a \code{knitr} context
}
