% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.vectors.R
\name{mean.vectors}
\alias{mean.vectors}
\title{\code{poisson.potential} returns an sf}
\usage{
\method{mean}{vectors}(pts, OD)
}
\description{
\code{poisson.potential} returns an sf
}
\examples{
# data(munich)
# sf=mean.vectors(munich$coords,minich$OD)
# ggplot(sf)+geom_segment(aes(x=x,y=y,xend=x+dx*100,yend=y+dy*100),arrow = arrow(angle=15,length = unit(0.3,"cm"),type = "closed"))+theme_void()
}
