% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_counts}
\alias{normalize_counts}
\title{this function is a wrapper for sctransform::vst}
\usage{
normalize_counts(counts)
}
\arguments{
\item{counts}{\itemize{
\item non-negative numeric matrix containing gene counts,
rows correspond to spots, columns correspond to genes
}}
}
\value{
normalized count matrix in the same format as parameter counts
}
\description{
this function is a wrapper for sctransform::vst
}
