% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adjustment.R
\name{plot_adjustment}
\alias{plot_adjustment}
\title{This function helps choosing offset and spacing parameters for good spatial plots. Basically just a wrapper for spatial_plot that returns the input parameters in a list for easy use later on.}
\usage{
plot_adjustment(ids, img, nx = 35, ny = 33, ox = 0, oy = 0)
}
\arguments{
\item{ids}{data frame or matrix assigning spatial coordinates to the spots (should be complete ids data frame, no missing spots)}

\item{img}{Image of the ST slide, loaded by EBImage::readImage()}

\item{nx}{numeric > 0, defines spacing between plotted spots on x-axis;
rule of thumb: number of spots shown in the image (horizontal); the large this number, the closer together the spots are}

\item{ny}{numeric > 0, defines spacing between plotted spots on y-axis;
rule of thumb: number of spots shown in the image (vertical); the larger this number, the closer together the spots are}

\item{ox}{numeric, defines the offset of spot 0,0 from the lower right corner of the image in x-direction (can be negative)}

\item{oy}{numeric, defines the offset of spot 0,0 from the lower tight corner of the image in y-direction (can be negative)}
}
\value{
list containing the input parameters nx, ny, ox, oy. If the produced plot looks good, this can be used to specify the parameters for plotting in other functions of this package.
}
\description{
This function helps choosing offset and spacing parameters for good spatial plots. Basically just a wrapper for spatial_plot that returns the input parameters in a list for easy use later on.
}
