% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_clustering.R
\name{plot_enrichment_clustering}
\alias{plot_enrichment_clustering}
\title{create plots for visualizing the enrichment and clustering results}
\usage{
plot_enrichment_clustering(
  counts,
  ids,
  img,
  clustering.list,
  filepath,
  filename,
  spatial.file.prefix = NULL,
  spot.cluster.names = NULL
)
}
\arguments{
\item{counts}{non-negative numeric matrix. rows correspond to spots,
columns correspond to genes}

\item{ids}{data frame with rownames corresponding to barcodes / spot names, assigning
spatial coordinates to each spot; coordinates are contained in columns X and Y and conform to
the same standards as the output of process_input()}

\item{img}{EBImage image object the information will be plotted on; default NULL}

\item{clustering.list}{output of cluster_spots}

\item{filepath}{directory in which results should be stored}

\item{filename}{character, filename for the clustering and enrichment plots; only filename, the path is given in filepath}

\item{spatial.file.prefix}{character, filename_prefix for several PDFs containing spatial enrichment plots; will be store in filepath/}

\item{spot.cluster.names}{character vector assigning each cluster in clustering.list$spot.clustering a name; for better visualization
(only really useful on second run, when clustering has been visualized and tissue can be named properly). Ordering of the names needs to
match a numerical ordering of the clusters.}
}
\value{
NULL, plots to files
}
\description{
create plots for visualizing the enrichment and clustering results
}
