% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichments.R
\name{plot_enrichments}
\alias{plot_enrichments}
\title{plot spatial heatmaps of enriched terms to pdf}
\usage{
plot_enrichments(
  enrichments,
  ids,
  counts,
  clustering,
  img = NULL,
  plot.params = list(nx = 35, ny = 33, ox = 0, oy = 0),
  filepath = NULL
)
}
\arguments{
\item{enrichments}{output of test_enrichment() function}

\item{ids}{barcode data frame assigning spatial coordinates to all spots; column names 'X' and 'Y'}

\item{counts}{non-negative numeric matrix containing gene expression for all spots (spots x genes)}

\item{clustering}{numeric vector assigning clusters to all spots. same order as spots in counts}

\item{img}{ST image (EBImage object) if available. Will be used as plot background. If no image is available, leave default NULL}

\item{plot.params}{list as created by plot_adjustment()}

\item{filepath}{character, directory in which the output file should be stored. If not specified, currecnt working directory is used.}
}
\value{
plot to file "enrichment_spatial.pdf" in 'filepath'. Returns the data frame containing the enrichment p-values for each term for all clusters.
}
\description{
plot spatial heatmaps of enriched terms to pdf
}
