% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spot_enrichments.R
\name{plot_spot_enrichments}
\alias{plot_spot_enrichments}
\title{this function plot spatial patterns of enriched GO terms to pdf}
\usage{
plot_spot_enrichments(
  enrichment.mat,
  term.names = NULL,
  ids,
  plot.params = list(nx = 35, ny = 33, ox = 0, oy = 0),
  img = NULL,
  filename = "spot_enrichments.pdf",
  output.path = "./"
)
}
\arguments{
\item{enrichment.mat}{numeric matrix of enrichment p-values (terms x spots) as returned by test_spot_enrichment or filter_spot_enrichments}

\item{term.names}{character vector containing full term names for GO ids in enrichment.mat}

\item{ids}{barcode data frame assigning spots to x and y coordinates (see process_input)}

\item{plot.params}{list of parameters needed for good spatial visualization as returned by plot_adjustment}

\item{img}{EBImage image object the information will be plotted on; default NULL}

\item{filename}{character, name of the pdf; default "spot_enrichments.pdf"}

\item{output.path}{character specifying output directory; default "./"}
}
\description{
this function plot spatial patterns of enriched GO terms to pdf
}
