% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_vintage.R
\name{plus_minus_map}
\alias{plus_minus_map}
\title{Create a +- map !}
\usage{
plus_minus_map(xsf, varname)
}
\arguments{
\item{xsf}{An \code{\link{sf}} object with poylgones or points}

\item{varname}{The name of the column that store the balances to plot}
}
\value{
a plot
}
\description{
Create a +- map !
}
\examples{
data(dollars)
vintage_frame(dollars$polygones)
dollars$polygones$delta = rowSums(dollars$OD)-colSums(dollars$OD)
plus_minus_map(dollars$polygones,"delta")
}
