% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.potential.R
\name{poisson.potential}
\alias{poisson.potential}
\title{\code{poisson.potential} returns an sf}
\usage{
poisson.potential(
  xsf,
  varname = "delta",
  method = "jacobi",
  nb_it = 1000,
  cellsize = sqrt((sf::st_bbox(xsf)[3] - sf::st_bbox(xsf)[1]) * (sf::st_bbox(xsf)[4] -
    sf::st_bbox(xsf)[2])/2500)
)
}
\arguments{
\item{xsf}{an sf data.frame of polygons}

\item{varname}{name of the feature where the balance of each polygons are stored}

\item{method}{resolution method for poisson equation either jacobi or solve}

\item{nb_it}{number of iteration for jacobi resolution}

\item{cellsize}{size of the}
}
\value{
an sf data.frame with
}
\description{
\code{poisson.potential} returns an sf
}
