% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_compute_poisson_potential.R
\name{poisson_flows_map}
\alias{poisson_flows_map}
\title{Utils to show the flows derived from a poisson potential.}
\usage{
poisson_flows_map(sf, normfact = 0.1)
}
\arguments{
\item{normfact}{a scalar (normalization factor for arrows length)}

\item{xsf}{an \code{sf} data.frame as returned by \code{compute_poisson_potential}}
}
\value{
an arrow plot
}
\description{
Utils to show the flows derived from a poisson potential.
}
