% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_lasso_data.R
\name{prepare_lasso_data}
\alias{prepare_lasso_data}
\title{process the lasso counts and ids for conformity with pipeline output
match spot names in raw ids and lasso ids}
\usage{
prepare_lasso_data(ids_raw, ids_lasso, counts_lasso)
}
\arguments{
\item{ids_raw}{data frame containing spatial information as built by
load_ids from and ST barcode file}

\item{ids_lasso}{data frame containing ids of lasso spots. For this the
order in which spots are stored in the lasso models needs to be known:
idx = i*x_max + j; 0 <= i <= x_max-1, 1 <= j <= y_max}

\item{counts_lasso}{the model that is built in build_lassos,
containing 34*32 spots without rownames}
}
\value{
list containing two elements\cr
\enumerate{
\item counts - the count matrix with rownames and reduced to original spots
\item ids - the lasso ids
}
}
\description{
after model building the resulting counts matrix and the
ids data frame do not contain rownames (spot names / spatial information).
This function matches spatial positions of lasso and raw ids and
integrates the information into the lasso model.
}
