% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_input.R
\name{process_input}
\alias{process_input}
\title{process input data from ST pipeline to make sure they conform
to the format the package expects}
\usage{
process_input(counts, ids = NULL, separate_by = "x", force_counts = FALSE)
}
\arguments{
\item{counts}{non-negative numeric matrix containing gene counts,
rows correspond to spots, columns correspond to genes}

\item{ids}{data frame or matrix assigning spatial coordinates to the spots,
the spot 0,0 is in the right lower corner of the
corresponding tissue slide and the number of spots along the X-axis is
higher than the number of spots on the Y-axis.
optional if spatial information is contained in rownames of counts}

\item{separate_by}{only necessary if spatial information is given as
in the output of ST Pipeline (Navarro et al.), e.g. 32x2; default x}

\item{force_counts}{logical, in case there are less genes than spots this needs to be TRUE;
default FALSE}

\item{force_indices}{logical, in case the range of x and y indices is not the same as in the
ST barcode files}
}
\value{
list with two entries:\cr
\enumerate{
\item counts - count matrix
\item ids - barcode data frame assigning spatial positions to spots
}
}
\description{
process input data from ST pipeline to make sure they conform
to the format the package expects
}
