% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_compute_poisson_potential.R
\name{rasterFromPotential}
\alias{rasterFromPotential}
\title{Utils to convert poisson potential to raster}
\usage{
rasterFromPotential(pot)
}
\arguments{
\item{pot}{an \code{sf} data.frame as returned by \code{compute_poisson_potential}}
}
\value{
a \code{raster}  representation of the potential
}
\description{
Utils to convert poisson potential to raster
}
