% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{this function is a wrapper for EBImage function readImage
read ST image and scale to 1000 x 1000}
\usage{
read_image(path)
}
\arguments{
\item{path}{filepath to ST slide image}
}
\description{
this function is a wrapper for EBImage function readImage
read ST image and scale to 1000 x 1000
}
