% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_term_mat.R
\name{reduce_term_mat}
\alias{reduce_term_mat}
\title{filter term matrix based on specificity and coverage}
\usage{
reduce_term_mat(
  term.mat,
  spot.cluster,
  exclusive.threshold = 0.75,
  cluster.marker = 0.85
)
}
\arguments{
\item{term.mat}{matrix containing enrichment information, terms x spots. Term not enriched in a spot have entry 0 in that column.}

\item{spot.cluster}{vector containing clustering information for the spots in term.mat}

\item{exclusive.threshold}{numeric >0, <1; if more than this fraction of spots in which a term is enriched are in the same cluster,
the term is kept}

\item{cluster.marker}{numeric >0, <1; if more than this fraction of the spots of any cluster are enriched for a certain term,
that term is kept}
}
\value{
term.mat, filtered for interesting terms
}
\description{
filter term matrix based on specificity and coverage
}
