% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_background.R
\name{remove_background}
\alias{remove_background}
\title{remove background spots (not covered by tissue) from ST data based on tissue image.
Manual checking of the result and adjustment of the threshold is needed.}
\usage{
remove_background(img, nx = 35, ny = 33, ids, names, threshold = 0.7)
}
\arguments{
\item{img}{image of the ST slide, cropped to the spot area. image read by EBimage::readImage()}

\item{nx}{number of spots shown in the image on the horizontal axis}

\item{ny}{number of spots shown in the image on the vertical axis}

\item{ids}{data frame assigning barcodes / spot names to spatial coordinates}

\item{names}{barcodes / names of the spots in the order in which they are stored in the counts matrix}

\item{threshold}{relative brightness. spots above this threshold are discarded}
}
\value{
list containing two entries\cr
\enumerate{
\item spots.to.keep - vector of barcodes / spot names of the spots that should not be removed
\item image - image showing the areas that are retained.
typically framed by spots for which there exist no measurements. Depending on whether these are contained
in the image or not, different values need to be chosen for nx, ny. If they are included, the
default values of nx=35 and ny=33 are sensible. If not, it should be nx=33 and ny=31.\cr
The image is assumed to be oriented such that the rectangular 4x4 array of spots in one of the corners
is in the lower right corner and on the x-axis (left to right) there are more spots than on the y-axis.
}
}
\description{
remove background spots (not covered by tissue) from ST data based on tissue image.
Manual checking of the result and adjustment of the threshold is needed.
}
\details{
nx and ny depend on the cropping of the image. the area of measurements in an ST experiment is
}
