% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_fs.R
\name{show_ttt_files}
\alias{show_ttt_files}
\title{Get a vector of files in ttt_dir}
\usage{
show_ttt_files(full_names = FALSE)
}
\arguments{
\item{full_names}{an option for showing the
directory paths prepended to the file names.
full path}
}
\description{
Get a vector of all files in the ttt directory.
}
