% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_orientation.R
\name{slide_orientation}
\alias{slide_orientation}
\title{this function changes the orientation of the data by
mirroring x or y coordinates and/or rotating to the right
by multiples of 90 degrees}
\usage{
slide_orientation(ids, flip.x = FALSE, flip.y = FALSE, rotate = 0)
}
\arguments{
\item{ids}{barcode data frame as returned by process_input}

\item{flip.x}{logical, should x coordinates be flipped? default FALSE}

\item{flip.y}{logical, should y coordinates be flipped? default FALSE}

\item{rotate}{integer, multiple of 90; degrees by which data is rotated to the right}
}
\value{
barcode data frame
}
\description{
this function changes the orientation of the data by
mirroring x or y coordinates and/or rotating to the right
by multiples of 90 degrees
}
\details{
If more than one action is specified, the tasks are carried out in the following order:\cr
\itemize{
\item flip x\cr
\item flip y\cr
\item rotate\cr
}
}
