% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_gene_plots.R
\name{spatial_gene_plots}
\alias{spatial_gene_plots}
\title{create spatial plots for the top n.genes genes in a given table and store them as pdf}
\usage{
spatial_gene_plots(
  gene.table,
  counts,
  ids,
  img = NULL,
  plot.params = NULL,
  n.genes = 100,
  filepath = NULL,
  colour = "black"
)
}
\arguments{
\item{gene.table}{data frame containing interesting genes, as returned by filter_genes (or entry differential.genes of analyze_clustering output)}

\item{counts}{non-negative numeric matrix containing gene expression for all spots (spots x genes)}

\item{ids}{barcode data frame assigning spatial coordinates to all spots; column names 'X' and 'Y'}

\item{img}{ST image (EBImage object) if available. Will be used as plotting background. If not available, leave default NULL.}

\item{plot.params}{list created by plot_adjustment; only needed if img is provided}

\item{n.genes}{integer, number of genes to be plotted (starting from the top of gene.table)}

\item{filepath}{character specifying the directory in which the output file ('spatial_genes.pdf') should be stored}

\item{colour}{character specifying the colour of the spots in the gene expression plots; default "black"}
}
\value{
NULL, plots to file 'filepath'/spatial_genes.pdf
}
\description{
create spatial plots for the top n.genes genes in a given table and store them as pdf
}
