% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_enrichment.R
\name{test_enrichment}
\alias{test_enrichment}
\title{perform enrichment testing based on differential gene expression analysis}
\usage{
test_enrichment(
  analysis.data,
  clustering,
  counts,
  split_up_down = TRUE,
  db_dataset = "mmusculus_gene_ensembl",
  exclusive = F,
  gene_id = "hgnc"
)
}
\arguments{
\item{analysis.data}{output of analyze_clustering}

\item{clustering}{numeric vector assigning clusters to all spots (same order as spots in counts matrix)}

\item{counts}{numeric non-negativ matrix containing gene expression for all spots (spots x genes)}

\item{split_up_down}{boolean, should enrichment testing be performed separately on the list of up and downregulated genes for each cluster? default TRUE}

\item{db_dataset}{ensembl dataset to use. default "mmusculus_gene_ensembl" (mouse), for human data use "hsapiens_gene_ensembl"}

\item{exclusive}{boolean, should each gene only be considered for enrichment analysis in the cluster where it has the lowest p-value? default FALSE}

\item{gene_id}{character denoting the gene symbols to use; default "hgnc" for HGNC symbols; use "ensembl" for ensembl gene ids}
}
\value{
list containing enrichmnet information for each cluster. if split_up_down is TRUE,
the list entry for each cluster is a list with two elements:\cr
\enumerate{
\item up - table with enrichment information for upregulated genes (as returned by topGO)\cr
\item down - table with enrichment information for downregulated genes (as returned by topGO)\cr
If split_up_down is false, the list contains one enrichment table for each cluster.
}
}
\description{
perform enrichment testing based on differential gene expression analysis
}
