% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_spot_enrichment.R
\name{test_spot_enrichment}
\alias{test_spot_enrichment}
\title{perform enrichment testing based on differential gene expression analysis}
\usage{
test_spot_enrichment(
  analysis.data,
  counts,
  db_dataset = "mmusculus_gene_ensembl",
  gene_id = "hgnc",
  ncores = 4,
  mirror = NULL
)
}
\arguments{
\item{analysis.data}{output of analyze_clustering}

\item{counts}{numeric non-negativ matrix containing gene expression for all spots (spots x genes)}

\item{db_dataset}{ensembl dataset to use. default "mmusculus_gene_ensembl" (mouse), for human data use "hsapiens_gene_ensembl"}

\item{gene_id}{character denoting the gene symbols to use; default "hgnc" for HGNC symbols; use "ensembl" for ensembl gene ids}

\item{ncores}{integer > 0, number of cores to use}

\item{mirror}{string indicating ensembl mirror site; must be one of "standard", "uswest", "useast", "asia"; default is "standard"}
}
\value{
list containing two entries:\cr
\enumerate{
\item enrichments - matrix contining p-values of enrichments test for
each spot and GO term enriched in at least one spot(terms x spots) \cr
\item term.names - character vector containing full names of the GO ids in the enrichment matrix
}
}
\description{
perform enrichment testing based on differential gene expression analysis
}
