% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{tt}
\alias{tt}
\title{Track Task (*tt*)}
\usage{
tt(..., task = NULL, s = NULL, e = NULL, lbl = NULL, proj = NULL,
  info = NULL, write = TRUE)
}
\arguments{
\item{...}{a string that allows for a shorthand means to
track time for some task.}

\item{task}{the task description.}

\item{s}{the start time for the task. Must be
in the ISO date/time format of \code{YYYY-MM-DD HH:MM}.}

\item{e}{the end time for the task. Must be
in the ISO date/time format of \code{YYYY-MM-DD HH:MM}.}

\item{lbl}{a label to use for the event.}

\item{proj}{project information for the event.}

\item{info}{information pertaining to the event.}

\item{write}{a logical value indicating whether the
tracked task entry should be written to disk.}
}
\description{
Track a task by using \code{tt()}.
}
\examples{
\dontrun{
# Add a tracked task using `tt()`; it will
# be committed to the ttt_dir; information
# on the tracked task will be echoed back
tt(
  task = "Description of the task",
  s = "2018-05-24 12:30",
  e = "2018-05-24 13:00",
  lbl = "label",
  proj = "project_name",
  info = "Any additional information")
#> tracked task entry
#> ------------------
#> t:  Description of the task
#> s:  2018-05-24 11:30
#> e:  2018-05-24 12:00
#> l:  label
#> p:  project_name
#> i:  Any additional information

# We can also use the shorthand notation
tt("
t: Description of the task;
s: 2018-05-24 11:30;
e: 2018-05-24 12:00;
l: label;
p: project_name;
i: Any additional information
")
#> tracked task entry
#> ------------------
#> t:  Description of the task
#> s:  2018-05-24 11:30
#> e:  2018-05-24 12:00
#> l:  label
#> p:  project_name
#> i:  Any additional information
}
}
