% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttd.R
\name{ttd}
\alias{ttd}
\title{Draft a Tracked Task}
\usage{
ttd(..., task = NULL, s = NULL, e = NULL, lbl = NULL, proj = NULL,
  info = NULL)
}
\arguments{
\item{...}{a string that allows for a shorthand means to
track time for some task.}

\item{task}{the task description.}

\item{s}{the start time for the task. Must be
in the ISO date/time format of \code{YYYY-MM-DD HH:MM}.}

\item{e}{the end time for the task. Must be
in the ISO date/time format of \code{YYYY-MM-DD HH:MM}.}

\item{lbl}{a label to use for the event.}

\item{proj}{project information for the event.}

\item{info}{information pertaining to the event.}
}
\description{
Use the syntax of the \code{tt()} function without
actually committing the task. This is useful for
practicing the shorthand of entering a tracked task
since the tracked-time event is echoed to the console.
}
