\name{ttt-package}
\alias{ttt-package}
\alias{ttt}
\docType{package}
\title{
\packageTitle{ttt}
}
\description{
\packageDescription{ttt}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ttt}
\packageIndices{ttt}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ttt}

Maintainer: \packageMaintainer{ttt}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
