% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_flowmapper_types.R
\name{ttt_flowmapper_types}
\alias{ttt_flowmapper_types}
\title{ttt_flowmapper_types: Compute gross and net flows}
\usage{
ttt_flowmapper_types(
  tab,
  origin = NULL,
  destination = NULL,
  fij = NULL,
  format,
  type,
  lowup,
  net
)
}
\arguments{
\item{tab}{is the input flow dataset}

\item{origin}{the place of origin code}

\item{destination}{the place of destination code}

\item{fij}{the flow value between origin and destination places}

\item{format}{specify the flow dataset format, "M " for square matrix or "L" for long [i,j,data]}

\item{type}{enter the flowtype : "gross" or "net". See Details.}

\item{lowup}{for extracting the lower or upper triangular sub-portion of the gross matrix. See Details.}

\item{net}{for extracting the "positive" or the "negative" flow values of the net matrix}
}
\description{
Compute gross (symmetric) and net (antisymetric) flows from initial asymetric matrix
}
\details{
Type of flow are:\cr 
- type = "gross" for building the gross matrix - as bilateral volum\cr
- type = "net" for building the net matrix - as bilateral balance\cr
Lowup part of the matrix are:\cr
- lowup ="up" for triangular part above the main diagonal \cr
- lowup = "low" for triangular part below the main diagonal\cr
}
\examples{
#Example 1 : building gross flows
gross_up<-ttt_flowmapper_types(data, origin="i",destination="j",fij="fij", format="L", type="gross",lowup="up")\cr
#Example 2 : building net flows
net_positive <-ttt_flowmapper_types(data,origin="i",destination="j",fij="fij",format="L",type="net",net="positive")\cr 
}
\references{
Waldo R. Tobler, 1979, A geographical flow mapping program, Geographical Analysis, n°13, Vol. 1, pp. 1-20
}
