% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_vintage.R
\name{vintage_frame}
\alias{vintage_frame}
\title{Create a vintage frame !}
\usage{
vintage_frame(x)
}
\arguments{
\item{x}{An \code{\link{sf}} object}
}
\value{
a plot
}
\description{
Create a vintage frame !
}
\examples{
data(dollars)
vintage_frame(dollars$polygones)
}
