% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_genes.R
\name{visualize_genes}
\alias{visualize_genes}
\title{This function visualizes the results of the clustering and analysis using spatial plots and heatmaps}
\usage{
visualize_genes(
  counts,
  ids,
  img,
  clustering,
  gene.info,
  filepath = NULL,
  plot.params = list(nx = 35, ny = 33, ox = -1000/70, oy = 1000/32)
)
}
\arguments{
\item{counts}{non-negative numeric matrix containing gene counts,
rows correspond to spots, columns correspond to genes}

\item{ids}{data frame or matrix assigning spatial coordinates to the spots}

\item{img}{image of ST slide, read using EBImage::readImage()}

\item{clustering}{numeric vector, same order as rownames(counts), assigning each spot a cluster}

\item{gene.info}{list as returned by analyze_clustering()}

\item{filepath}{character, directory in which to save plots; if not specified, plots will not be saved to disk}

\item{plot.params}{plotting parameters as returned by plot_adjustment()}
}
\value{
list containing 4 plots (ggplot objects):\cr
\enumerate{
\item spatial.cluster - overlay of clustering on ST image
\item spatial.cluster.legend - spatial clustering, without ST image in background, but with legend
\item heatmap.plt - simplified heatmap, depicting mean expression of all interesting genes in the different clusters
\item heatmap.full.plt - heatmap visualizing expression of all interesting genes in all spots
}
}
\description{
This function visualizes the results of the clustering and analysis using spatial plots and heatmaps
}
