% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_compute_poisson_potential.R
\name{winds}
\alias{winds}
\title{animated droplets map of poisson potentials results 
winds(pot,nbparticules = 1000,lifespan = 50,resolution = 0.2,scalefact = 0.1)}
\usage{
winds(
  pot,
  lifespan = 5,
  scalefact = 0.1,
  linewidth = 1,
  nbparticules = 500,
  resolution = sf::st_bbox(pot[1, ])[3] - sf::st_bbox(pot[1, ])[1]
)
}
\description{
animated droplets map of poisson potentials results 
winds(pot,nbparticules = 1000,lifespan = 50,resolution = 0.2,scalefact = 0.1)
}
