% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_clustering.R
\name{analyze_clustering}
\alias{analyze_clustering}
\title{This function analyzes ST data based on clustering generated by cluster_counts_OL.
It calculates a list of genes specific to a certain cluster and identifies differentually expressed genes
across clusters}
\usage{
analyze_clustering(counts, ids, clustering, sig.level = 0.05)
}
\arguments{
\item{counts}{non-negative numeric matrix containing gene counts,
rows correspond to spots, columns correspond to genes}

\item{ids}{data frame or matrix assigning spatial coordinates to the spots (see process_input() for details)}

\item{clustering}{list containing clustering information as returned by cluster_counts_OL (numeric vector assigning spots to clusters)}

\item{sig.level}{significance level for the t-test between clusters; genes with p-values above this threshold
will be removed from output; default 0.05}
}
\value{
list with 3 entries:\cr
\enumerate{
\item specific - character vecor containing names of genes that are specific to any cluster\cr
\item differential genes - data frame containing information (name, ckuster, p-value, up-/downregulation) for differentially expressed genes\cr
\item dsg - differential genes reduced to genes that are also contained in 'specific'
}
}
\description{
This function analyzes ST data based on clustering generated by cluster_counts_OL.
It calculates a list of genes specific to a certain cluster and identifies differentually expressed genes
across clusters
}
