% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lassos.R
\name{build_lassos}
\alias{build_lassos}
\title{Calculate 2D fused-lasso solutionpath for a countmatrix and save as object}
\usage{
build_lassos(
  counts_matrix,
  coords_table,
  name,
  output_folder = NULL,
  ncores = 4,
  gamma = 1
)
}
\arguments{
\item{counts_matrix}{input matrix of counts with genes in columns and spot identifier in rows}

\item{coords_table}{coordinate table which converts the spot identifier into an x and y coordinate on a grid}

\item{name}{Name of the experiment}

\item{output_folder}{folder in which to save the results}

\item{ncores}{number of parallel processes to start (default 8)}

\item{gamma}{ratio of penalization between sparsity and continuity (default 1, high = sparse)}
}
\value{
list containing two elements\cr
\enumerate{
\item counts - count matrix in the same format as input count matrix
\item ids - id data frame that assigns spot identifiers to locations on the grid
}
}
\description{
Calculate 2D fused-lasso solutionpath for a countmatrix and save as object
}
\details{
This function builds a solution path for each gene (column) in a input matrix, selects the models according to the BIC and
builds a count matrix
}
