% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BICs.R
\name{calc_BICs}
\alias{calc_BICs}
\title{Returns useful information about a fusedlasso model fit}
\usage{
calc_BICs(full_model)
}
\arguments{
\item{full_model}{takes an object of the class fusedlasso}
}
\value{
returns a vector consisting of the df, lambda, rss, BIC and ratio of explained variation dor each lambda.
}
\description{
Returns useful information about a fusedlasso model fit
}
\details{
This function calculates the RSS, BIC and explained variance of a genlasso model
}
