% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_genes}
\alias{filter_genes}
\title{this function returns a table of interesting genes based on lasso model fit, specificity and differential expression}
\usage{
filter_genes(analysis.data, lasso.data)
}
\arguments{
\item{analysis.data}{output of analyze_clustering function}

\item{lasso.data}{output of build_lassos function}
}
\value{
data frame containing gene name, cluster with lowest p-value, lowest p-value and up/downregulation information for each gene
}
\description{
this function returns a table of interesting genes based on lasso model fit, specificity and differential expression
}
