% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt.R
\name{print.ttt}
\alias{print.ttt}
\title{Print \code{ttt} object}
\usage{
\method{print}{ttt}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{ttt}}.}

\item{...}{Further arguments passed on to other \code{print} methods.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print \code{ttt} object
}
\details{
In an interactive context, the rendered table will be displayed in
a web browser. Otherwise, the HTML code will be printed as text.
}
